+++ CHANGES.rst
@@ -10,6 +10,10 @@ https://github.com/zopefoundation/Zope/blob/4.x/CHANGES.rst
 5.0a1 (unreleased)
 ------------------
 
+- Fix longstanding test bug by forcing the page template engine.
+  Many tests in ``Products.PageTemplates`` used the old Zope page template
+  engine because the correct one was not registered during setup.
+
 - Remove deprecated module ``ZPublisher.maybe_lock``
   (`#758 <https://github.com/zopefoundation/Zope/issues/758>`_)